+++ src/gate.ts
@@ -197,6 +197,7 @@ export class Gate {
     }
     // handle connection failures
     handleFailure(failure: Failure) {
+        this.t7.log(failure)
         if (!this.boarding)
             return
 
@@ -220,7 +221,7 @@ export class Gate {
         }
         if (failure == Failure.TimedOut) {
             if ((!this.fp) && this.tryWebexec) {
-                this.notify("webexec server timed out. Trying SSH...")
+                this.notify("Timed out, trying SSH...")
                 this.tryWebexec = false
                 this.connect()
                 return
@@ -244,7 +245,7 @@ export class Gate {
             return
         // if we're already boarding, just focus
         if (this.boarding) {
-            console.log("already boarding")
+            this.t7.log("already boarding")
             if (!this.windows || (this.windows.length == 0))
                 this.activeW = this.addWindow("", true)
             this.focus()
@@ -274,7 +275,11 @@ export class Gate {
     }
     // reset reset's a gate connection by disengaging and reconnecting
     reset() {
-        this.disengage().then(() => this.t7.run(() => this.connect(), 100))
+        this.disengage().then(() => {
+            this.t7.run(() =>  {
+                this.connect()
+            }, 100)
+        })
     }
     async loseState () {
         const fp = await this.t7.getFingerprint(),
@@ -314,7 +319,7 @@ echo "${fp}" >> ~/.config/webexec/authorized_fingerprints
             this.t7.log("Fresh state, creating the first pane")
             if (winLen > 0)
                 // TODO: find a way to identify state lost
-                console.log("this.loseState()")
+                this.t7.log("this.loseState()")
             else
                 this.activeW = this.addWindow("", true)
         } else if (winLen > 0) {
@@ -364,7 +369,7 @@ echo "${fp}" >> ~/.config/webexec/authorized_fingerprints
      * clear clears the gates memory and display
      */
     clear() {
-        console.log("Clearing gate")
+        this.t7.log("Clearing gate")
         this.e.querySelector(".tabbar-names").innerHTML = ""
         this.e.querySelectorAll(".window").forEach(e => e.remove())
         this.e.querySelectorAll(".modal").forEach(e => e.classList.add("hidden"))
@@ -413,7 +418,7 @@ echo "${fp}" >> ~/.config/webexec/authorized_fingerprints
                this.sendStateTask = null
                this.session.setPayload(this.dump()).then(() => {
                     if ((this.windows.length == 0) && (this.session != null)) {
-                        console.log("Closing gate after updating to empty state")
+                        this.t7.log("Closing gate after updating to empty state")
                         this.disengage().then(() => {
                             this.session = null
                             this.stopBoarding()
@@ -572,7 +577,7 @@ echo "${fp}" >> ~/.config/webexec/authorized_fingerprints
             this.notify("TBD: update new layout")
             this.t7.log("TBD: update layout", layout)
         }
-        console.log("opening session")
+        this.t7.log("opening session")
         this.session.connect()
     }
 }